# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "goose-server",
    "description": "An AI agent",
    "contact": {
      "name": "Block",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.30.0"
  },
  "paths": {
...
    "/sessions/{id}/cancel": {
      "post": {
        "tags": [
          "super::routes::session_events"
        ],
        "operationId": "session_cancel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Cancellation accepted"
          }
        }
      }
    },
    "/sessions/{id}/events": {
      "get": {
        "tags": [
          "super::routes::session_events"
        ],
        "operationId": "session_events",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE event stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/MessageEvent"
                }
              }
            }
          },
          "404": {
            "description": "Session not found"
          }
        }
      }
    },
    "/sessions/{id}/reply": {
      "post": {
        "tags": [
          "super::routes::session_events"
        ],
        "operationId": "session_reply",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionReplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Request accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionReplyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Session not found"
          },
          "424": {
            "description": "Agent not initialized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/sessions/{session_id}": {
      "get": {
        "tags": [
          "Session Management"
        ],
        "operationId": "get_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Unique identifier for the session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Session Management"
        ],
        "operationId": "delete_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Unique identifier for the session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session deleted successfully"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CancelRequest": {
        "type": "object",
        "required": [
          "request_id"
        ],
        "properties": {
          "request_id": {
            "type": "string"
          }
        }
      },
...
      "Conversation": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Message"
        }
      },
...
      "ExtensionData": {
        "type": "object",
        "description": "Extension data containing all extension states\nKeys are in format \"extension_name.version\" (e.g., \"todo.v0\")",
        "additionalProperties": {}
      },
...
      "Message": {
        "type": "object",
        "description": "A message to or from an LLM",
        "required": [
          "role",
          "created",
          "content",
          "metadata"
        ],
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageContent"
            }
          },
          "created": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "$ref": "#/components/schemas/MessageMetadata"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
...
      "MessageEvent": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "message",
              "token_state",
              "type"
            ],
            "properties": {
              "message": {
                "$ref": "#/components/schemas/Message"
              },
              "token_state": {
                "$ref": "#/components/schemas/TokenState"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Message"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "error",
              "type"
            ],
            "properties": {
              "error": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Error"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "reason",
              "token_state",
              "type"
            ],
            "properties": {
              "reason": {
                "type": "string"
              },
              "token_state": {
                "$ref": "#/components/schemas/TokenState"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Finish"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "request_id",
              "message",
              "type"
            ],
            "properties": {
              "message": {
                "type": "object"
              },
              "request_id": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "Notification"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "conversation",
              "type"
            ],
            "properties": {
              "conversation": {
                "$ref": "#/components/schemas/Conversation"
              },
              "type": {
                "type": "string",
                "enum": [
                  "UpdateConversation"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Sent at the start of an SSE stream to inform the client about\nin-flight requests it can reattach to.",
            "required": [
              "request_ids",
              "type"
            ],
            "properties": {
              "request_ids": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "type": {
                "type": "string",
                "enum": [
                  "ActiveRequests"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "Ping"
                ]
              }
            }
          }
        ],
        "discriminator": {
          "propertyName": "type"
        }
      },
...
      "Session": {
        "type": "object",
        "required": [
          "id",
          "working_dir",
          "name",
          "created_at",
          "updated_at",
          "extension_data",
          "message_count"
        ],
        "properties": {
          "accumulated_input_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "accumulated_output_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "accumulated_total_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "conversation": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Conversation"
              }
            ],
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "extension_data": {
            "$ref": "#/components/schemas/ExtensionData"
          },
          "goose_mode": {
            "$ref": "#/components/schemas/GooseMode"
          },
          "id": {
            "type": "string"
          },
          "input_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "message_count": {
            "type": "integer",
            "minimum": 0
          },
          "model_config": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ModelConfig"
              }
            ],
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "output_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "provider_name": {
            "type": "string",
            "nullable": true
          },
          "recipe": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Recipe"
              }
            ],
            "nullable": true
          },
          "schedule_id": {
            "type": "string",
            "nullable": true
          },
          "session_type": {
            "$ref": "#/components/schemas/SessionType"
          },
          "total_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_recipe_values": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "user_set_name": {
            "type": "boolean"
          },
          "working_dir": {
            "type": "string"
          }
        }
      },
...
      "SessionReplyRequest": {
        "type": "object",
        "required": [
          "request_id",
          "user_message"
        ],
        "properties": {
          "override_conversation": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "nullable": true
          },
          "recipe_name": {
            "type": "string",
            "nullable": true
          },
          "recipe_version": {
            "type": "string",
            "nullable": true
          },
          "request_id": {
            "type": "string",
            "description": "Client-generated UUIDv7 identifying this request."
          },
          "user_message": {
            "$ref": "#/components/schemas/Message"
          }
        }
      },
      "SessionReplyResponse": {
        "type": "object",
        "required": [
          "request_id"
        ],
        "properties": {
          "request_id": {
            "type": "string"
          }
        }
      },
...
    }
  }
}